requirements.txt
# Required Python packages
streamlit
python-docx
sentence-transformers
chromadb
requests
transformers
sentencepiece
torch
